+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- (PA-6027) Add support for Ubuntu 24.04 noble
 
 ## [0.116.0] - 2024-02-23
 ### Added